Actualizacion al 28-04-2024

+++ ApiAuditoria/Models/AuditoriaAsignacion.cs
@@ -5,11 +5,11 @@ namespace ApiAuditoria.Models
     [Keyless]
     public class AuditoriaAsignacion
     {
-        public int aa_empresa {get; set;}
+        public Int16 aa_empresa {get; set;}
         public int aa_auditoria { get; set; }
-        public int aa_tarea { get; set; }
-        public int aa_secuencia { get; set; }
-        public int aa_responsable { get; set; }
+        public Int16 aa_tarea { get; set; }
+        public Int16 aa_secuencia { get; set; }
+        public Int16 aa_responsable { get; set; }
         public string? aa_tipo { get; set; }
         public string? aa_rol { get; set; }
         public string? aa_estado { get; set; }
