+++ ApiAuditoria/Models/AuditoriaGastos.cs
@@ -5,10 +5,10 @@ namespace ApiAuditoria.Models
     [Keyless]
     public class AuditoriaGastos
     {
-        public int ag_empresa {get; set;}
+        public Int16 ag_empresa {get; set;}
         public int ag_auditoria { get; set; }
-        public int ag_secuencia { get; set; }
-        public int ag_tipo { get; set; }
+        public Int16 ag_secuencia { get; set; }
+        public Int16 ag_tipo { get; set; }
         public DateTime ag_fecha_inicio { get; set; }
         public DateTime ag_fecha_fin { get; set; }
         public double ag_valor { get; set; }
