+++ ApiAuditoria/Models/AuditoriaTareaProcesos.cs
@@ -5,12 +5,12 @@ namespace ApiAuditoria.Models
     [Keyless]
     public class AuditoriaTareaProcesos
     {
-        public int at_empresa {get; set;}
+        public Int16 at_empresa {get; set;}
         public int at_auditoria { get; set; }
-        public int at_tarea { get; set; }
-        public int at_secuencia { get; set; }
-        public int at_auditor { get; set; }
-        public int at_responsable { get; set; }
+        public Int16 at_tarea { get; set; }
+        public Int16 at_secuencia { get; set; }
+        public Int16 at_auditor { get; set; }
+        public Int16 at_responsable { get; set; }
         public DateTime at_fecha { get; set; }
         public string? at_observaciones { get; set; }
         public string? at_estado { get; set; }
