+++ ApiAuditoria/Models/AuditoriaTareas.cs
@@ -5,10 +5,10 @@ namespace ApiAuditoria.Models
     [Keyless]
     public class AuditoriaTareas
     {
-        public int at_empresa {get; set;}
+        public Int16 at_empresa {get; set;}
         public int at_auditoria { get; set; }
-        public int at_tarea { get; set; }
-        public int at_oficina { get; set; }
+        public Int16 at_tarea { get; set; }
+        public Int16 at_oficina { get; set; }
         public string? at_asignacion { get; set; }
         public string? at_estado { get; set; }
         public string? at_usuario_creacion { get; set; }
