+++ ApiAuditoria/Models/CatalogoGastos.cs
@@ -5,8 +5,8 @@ namespace ApiAuditoria.Models
     [Keyless]
     public class CatalogoGastos
     {
-        public int cg_empresa {get; set;}
-        public int cg_codigo { get; set; }
+        public Int16 cg_empresa {get; set;}
+        public Int16 cg_codigo { get; set; }
         public string? cg_descripion { get; set; }
         public string? cg_estado { get; set; }
         public string? cg_usuario_creacion { get; set; }
