+++ ApiAuditoria/Models/CatalogoProcesos.cs
@@ -5,8 +5,8 @@ namespace ApiAuditoria.Models
     [Keyless]
     public class CatalogoProcesos
     {
-        public int cp_empresa {get; set;}
-        public int cp_codigo { get; set; }
+        public Int16 cp_empresa {get; set;}
+        public Int16 cp_codigo { get; set; }
         public string? cp_descripcion { get; set; }
         public string? cp_estado { get; set; }
         public string? cp_usuario_creacion { get; set; }
