+++ ApiAuditoria/Models/Monedas.cs
@@ -5,7 +5,7 @@ namespace ApiAuditoria.Models
     [Keyless]
     public class Monedas
     {
-        public int mo_codigo {get; set;}
+        public Int16 mo_codigo {get; set;}
         public string? mo_descripcion { get; set; }
         public string? mo_estado { get; set; }
         public string? mo_usuario_creacion { get; set; }
